+++ cmd/main.go
@@ -3,200 +3,17 @@ package main
 import (
 	"context"
 	"flag"
-	"fmt"
-	"io"
-	"lesson1/internal/config"
 	"log"
-	"net/http"
-	"net/url"
 	"os"
 	"os/signal"
-	"strings"
 	"sync"
-	"sync/atomic"
 	"syscall"
 	"time"
 
-	"github.com/PuerkitoBio/goquery"
+	"lesson1/internal/config"
+	"lesson1/internal/crawler"
+	"lesson1/internal/requester"
 )
-
-type CrawlResult struct {
-	Err   error
-	Info  string
-	Title string
-	Url   string
-}
-
-type Page interface {
-	GetTitle() string
-	GetLinks() []string
-	makeFullUrl(string) string
-}
-
-type page struct {
-	url     string
-	mainUrl string
-	doc     *goquery.Document
-}
-
-func NewPage(inUrl string, raw io.Reader) (Page, error) {
-	doc, err := goquery.NewDocumentFromReader(raw)
-	if err != nil {
-		return nil, err
-	}
-
-	//сохраняем основной url сайта
-	mainUrl := inUrl
-	u, err := url.Parse(inUrl)
-	if err == nil {
-		if u.User != nil {
-			mainUrl = fmt.Sprintf("%s://%s@%s", u.Scheme, u.User, u.Host)
-		} else {
-			mainUrl = fmt.Sprintf("%s://%s", u.Scheme, u.Host)
-		}
-	}
-	mainUrl = strings.TrimRight(mainUrl, "/")
-
-	return &page{url: inUrl, mainUrl: mainUrl, doc: doc}, nil
-}
-
-func (p *page) makeFullUrl(shortUrl string) string {
-	if strings.Index(shortUrl, "http") != 0 {
-		shortUrl = strings.TrimLeft(shortUrl, "/")
-		return p.mainUrl + "/" + shortUrl
-	}
-	return shortUrl
-}
-
-func (p *page) GetTitle() string {
-	return p.doc.Find("title").First().Text()
-}
-
-func (p *page) GetLinks() []string {
-	var urls []string
-	p.doc.Find("a").Each(func(_ int, s *goquery.Selection) {
-		url, ok := s.Attr("href")
-		if ok {
-			url = p.makeFullUrl(url)
-			urls = append(urls, url)
-		}
-	})
-	return urls
-}
-
-type Requester interface {
-	Get(ctx context.Context, url string) (Page, error)
-}
-
-type requester struct {
-	timeout time.Duration
-}
-
-func NewRequester(timeout time.Duration) requester {
-	return requester{timeout: timeout}
-}
-
-func (r requester) Get(ctx context.Context, url string) (Page, error) {
-	select {
-	case <-ctx.Done():
-		return nil, nil
-	default:
-		cl := &http.Client{
-			Timeout: r.timeout,
-		}
-		req, err := http.NewRequest("GET", url, nil)
-		if err != nil {
-			return nil, err
-		}
-		body, err := cl.Do(req)
-		if err != nil {
-			return nil, err
-		}
-		defer body.Body.Close()
-		page, err := NewPage(url, body.Body)
-		if err != nil {
-			return nil, err
-		}
-		return page, nil
-	}
-
-}
-
-//Crawler - интерфейс (контракт) краулера
-type Crawler interface {
-	Scan(ctx context.Context, wg *sync.WaitGroup, url string, depth int)
-	ChanResult() <-chan CrawlResult
-	ToChanResult(CrawlResult)
-	DepthDiff(int32)
-}
-
-type crawler struct {
-	r         Requester
-	res       chan CrawlResult
-	visited   map[string]struct{}
-	mu        sync.RWMutex
-	depthDiff int32 // для изменения depth
-}
-
-func NewCrawler(r Requester) *crawler {
-	return &crawler{
-		r:         r,
-		res:       make(chan CrawlResult),
-		visited:   make(map[string]struct{}),
-		mu:        sync.RWMutex{},
-		depthDiff: 0,
-	}
-}
-
-func (c *crawler) DepthDiff(diff int32) {
-	atomic.AddInt32(&c.depthDiff, diff)
-}
-
-func (c *crawler) Scan(ctx context.Context, wg *sync.WaitGroup, url string, depth int) {
-	defer wg.Done()
-
-	if depth <= 0 { //Проверяем то, что есть запас по глубине
-		return
-	}
-
-	c.mu.Lock()
-	_, ok := c.visited[url] //Проверяем, что мы ещё не смотрели эту страницу
-	if ok {
-		c.mu.Unlock()
-		return
-	}
-	c.visited[url] = struct{}{} //Помечаем страницу просмотренной (еще до того как посмотрели, чтоб другие грутины не пытались паралельно)
-	c.mu.Unlock()
-
-	select {
-	case <-ctx.Done(): //Если контекст завершен - прекращаем выполнение
-		return
-	default:
-		page, err := c.r.Get(ctx, url) //Запрашиваем страницу через Requester
-		if err != nil {
-			c.res <- CrawlResult{Err: err} //Записываем ошибку в канал
-			return
-		}
-		c.res <- CrawlResult{ //Отправляем результаты в канал
-			Title: page.GetTitle(),
-			Url:   url,
-		}
-		for _, link := range page.GetLinks() {
-			wg.Add(1)
-			c.mu.RLock()
-			go c.Scan(ctx, wg, link, int(c.depthDiff)+depth-1) //На все полученные ссылки запускаем новую рутину сборки
-			c.mu.RUnlock()
-		}
-	}
-}
-
-func (c *crawler) ChanResult() <-chan CrawlResult {
-	return c.res
-}
-
-func (c *crawler) ToChanResult(crawResult CrawlResult) {
-	c.res <- crawResult
-}
 
 func main() {
 
@@ -213,11 +30,11 @@ func main() {
 		panic(err)
 	}
 
-	var cr Crawler
-	var r Requester
+	var cr crawler.Crawler
+	var r requester.Requester
 
-	r = NewRequester(time.Duration(cfg.RequestTimeout) * time.Second)
-	cr = NewCrawler(r)
+	r = requester.NewRequester(time.Duration(cfg.RequestTimeout) * time.Second)
+	cr = crawler.NewCrawler(r)
 
 	ctx, cancel := context.WithTimeout(context.Background(), time.Duration(cfg.AppTimeout*int(time.Second)))
 	var wg sync.WaitGroup
@@ -226,7 +43,7 @@ func main() {
 	go processResult(ctx, cancel, cr, cfg)      //Обрабатываем результаты в отдельной рутине
 	go func() {
 		wg.Wait()
-		cr.ToChanResult(CrawlResult{
+		cr.ToChanResult(crawler.CrawlResult{
 			Info: "All urls already scanned", //Записываем сообщение канал
 		})
 		cancel() //все сканы завершились, а maxResult не достигнут
@@ -255,7 +72,7 @@ func main() {
 	}
 }
 
-func processResult(ctx context.Context, cancel func(), cr Crawler, cfg config.Config) {
+func processResult(ctx context.Context, cancel func(), cr crawler.Crawler, cfg config.Config) {
 	var maxResult, maxErrors = cfg.MaxResults, cfg.MaxErrors
 	for {
 		select {
